fix: cli structure
